Show Records count and total Records duration.

+++ app/src/main/java/com/dimowner/audiorecorder/v2/settings/SettingsState.kt
@@ -21,7 +21,7 @@ data class SettingsState(
     val recordingSettingsText: String,
     val rateAppLink: String,
     val feedbackEmail: String,
-    val totalRecordCount: Long,
+    val totalRecordCount: Int,
     val totalRecordDuration: Long,
     val availableSpace: Long,
     val appName: String,
